requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
trafilatura
beautifulsoup4
pdfminer.six
pdfplumber
openpyxl
pandas
python-docx
google-api-python-client
google-auth
tiktoken
langchain
langchain-text-splitters
python-multipart
groq
rank-bm25
chromadb
sentence-transformers
streamlit
sqlalchemy
PyMuPDF
